requirements.txt
Django==1.7.5
django-allauth==0.19.1
django-cors-headers==1.0.0
django-filter==0.9.2
django-oauth2-provider==0.2.6.1
django-rest-auth==0.3.4
djangorestframework==3.0.5
oauthlib==0.7.2
Pillow==2.7.0
psycopg2==2.6
python-openid==2.2.5
requests==2.5.3
requests-oauthlib==0.4.2
shortuuid==0.4.2
six==1.9.0
